==> notebooks/01_clean_1_month.ipynb <==
### Пополнување на вредностите за month

import pandas as pd

df = pd.read_csv('dataset/Spenser_1.tsv', sep='\t')
df

#### Пополнување на вредностите за month на местата каде што е извршена контрола.

Секој ден во дадена година според 4-4-5 форматот припаѓа на точно еден месец.

day = list(df.Calendarday)
month = list(df.month)

cnt = 0

for i in range(1, len(month)-1):
    if month[i] == ' ':
        if month[i-1] == month[i+1]:
            month[i] = month[i-1]
        else:
            cnt += 1

print(cnt)

calendar = dict()

for i in range(len(month)):
    if month[i] != ' ':
        calendar[day[i]] = month[i]

cnt = 0

for i in range(len(month)):
    if month[i] == ' ':
        if day[i] in calendar.keys():
            month[i] = calendar[day[i]]
        else:
            cnt += 1
            
print(cnt)

df.month = pd.Series(month)

df[df['month'] == ' ']

calendar

Проверка дали се пополнуваат точните полиња...

month[79841]

month[80936]

Во dataset-от овие вредности имаат index +2 од долунаведените.

Оваа вредност е пополнета со '5' бидејќи во dataset-от вредноста за FPE е иста со следната ставка (која има month = '5').

month[79841] = '5'

Оваа вредност е пополнета со '9' бидејќи двата соседни датуми во calendar-от ја имаат таа вредност.

month[80936] = '9'

Промена на типот на вредностите од string во int.

month_int = [int(month[i]) for i in range(len(month))]

set(month_int)

Замена на month во dataset-от и export на dataset-от по оваа промена.

df.month = pd.Series(month)

df.to_csv('dataset/Spenser_1_1_month.tsv', sep='\t')
