==> notebooks/01_clean_2_employee.ipynb <==
### Пополнување на вредностите за TERRID (Employee)

import pandas as pd

df = pd.read_csv('dataset/Spenser_1_1_month.tsv', sep='\t')
df

#### Пополнување на вредностите за TERRID на местата каде што е извршена контрола.

Секој customer припаѓа на точно еден employee.

cust_id = list(df.CustomerID)
terrid = list(df.TERRID)

cnt = 0

for i in range(1, len(terrid)-1):
    if terrid[i] == ' ':
        if terrid[i-1] == terrid[i+1]:
            terrid[i] = terrid[i-1]
        else:
            cnt += 1

print(cnt)

customer = dict()

for i in range(len(terrid)):
    if terrid[i] != ' ':
        customer[cust_id[i]] = terrid[i]

cnt = 0

for i in range(len(terrid)):
    if terrid[i] == ' ':
        if cust_id[i] in customer.keys():
            terrid[i] = customer[cust_id[i]]
        else:
            cnt += 1
            
print(cnt)

df.TERRID = pd.Series(terrid)

df[df['TERRID'] == ' ']

set(df[df['TERRID'] == ' '].index)

Начинот на пополнување јасно се гледа од dataset-от. Напишани се само дополнителни објаснувања. Во dataset-от овие вредности имаат index +2 од долунаведените.

terrid[41473] = 'Employee 60'
terrid[41474] = 'Employee 60'
terrid[41475] = 'Employee 60'
terrid[41476] = 'Employee 60'
terrid[41477] = 'Employee 60'
terrid[41478] = 'Employee 60'
terrid[140035] = 'Employee 49'

Проверка дали сите вредности се пополнети.

' ' in set(terrid)

Замена во dataset-от и export на dataset-от по оваа промена.

df.TERRID = pd.Series(terrid)

df.to_csv('dataset/Spenser_1_2_terrid.tsv', sep='\t')
