==> notebooks/01_clean_3_team.ipynb <==
### Пополнување на вредностите за SAPECCSalesGroup (Team)

import pandas as pd

df = pd.read_csv('dataset/Spenser_1_2_terrid.tsv', sep='\t')
df

#### Пополнување на вредностите за SAPECCSalesGroup на местата каде што е извршена контрола.

Секој employee припаѓа на точно еден team.

team = list(df.SAPECCSalesGroup)
terrid = list(df.TERRID)

cnt = 0

for i in range(1, len(team)-1):
    if team[i] == ' ':
        if team[i-1] == team[i+1]:
            team[i] = team[i-1]
        else:
            cnt += 1

print(cnt)

d = dict()

for i in range(len(team)):
    if team[i] != ' ':
        d[terrid[i]] = team[i]

cnt = 0

for i in range(len(team)):
    if team[i] == ' ':
        if terrid[i] in d.keys():
            team[i] = d[terrid[i]]
        else:
            cnt += 1
            
print(cnt)

df.SAPECCSalesGroup = pd.Series(team)

df[df['SAPECCSalesGroup'] == ' ']

set(df[df['SAPECCSalesGroup'] == ' '].index)

Пополнето со 'Team 6' бидејќи претходниот ('Employee 35') и следниот ('Employee 37') се од 'Team 6'. Во dataset-от овие вредности имаат index +2 од долунаведените.

team[138984] = 'Team 6'
team[138985] = 'Team 6'
team[138986] = 'Team 6'

Проверка дали сите вредности се пополнети.

' ' in set(team)

Замена во dataset-от и export на dataset-от по оваа промена.

df.SAPECCSalesGroup = pd.Series(team)

df.to_csv('dataset/Spenser_1_3_team.tsv', sep='\t')
