==> notebooks/01_clean_4_calendar.ipynb <==
### Замена на типот на Calendarday во pandas.Timestamp

import pandas as pd

df = pd.read_csv('dataset/Spenser_1_3_team.tsv', sep='\t')
df

datetime = pd.to_datetime(df.Calendarday)

df.Calendarday = datetime
df

Проверка дали сите вредности се точно променети.

incorrect = 0

for date in df.Calendarday:
    if type(date) is not pd._libs.tslib.Timestamp:
        incorrect += 1
        
print(incorrect)

Export на dataset-от по оваа промена.

df.to_csv('dataset/Spenser_1_4_calendar.tsv', sep='\t')
