==> notebooks/01_clean_5_standardize_coefficients.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('dataset/Spenser_1_4_calendar.tsv', sep='\t')

Стандардизирање на вредностите од SAPECoeficient атрибутот во формат '#.#'

sape_formatted = []

for i in range(len(df)):
    coefficient_str = df['SAPECoeficient'][i]
    if coefficient_str != ' ':
        coefficient = round(float(coefficient_str.replace(',','.')), 3)
        sape_formatted.append(coefficient)
    else:
        sape_formatted.append(coefficient_str)

df.SAPECoeficient = sape_formatted

Стандардизирање на вредностите од APECoeficient атрибутот во формат '#.#'

ape_formatted = []

for i in range(len(df)):
    coefficient_str = df['APECoeficient'][i]
    if coefficient_str != ' ':
        coefficient = round(float(coefficient_str.replace(',','.')), 3)
        ape_formatted.append(coefficient)
    else:
        ape_formatted.append(coefficient_str)

df.APECoeficient = ape_formatted

Стандардизирање на вредностите од FPECoeficient атрибутот во формат '#.#'

fpe_formatted = []

for i in range(len(df)):
    coefficient_str = df['FPECoeficient'][i]
    if coefficient_str != ' ':
        coefficient = round(float(coefficient_str.replace(',','.')), 3)
        fpe_formatted.append(coefficient)
    else:
        fpe_formatted.append(coefficient_str)

df.FPECoeficient = fpe_formatted

df.to_csv('dataset/Spenser_1_5_standardize_coefficients.tsv', sep='\t', index=None)
