==> notebooks/01_clean_6_SAPE_APE_FPE_string_to_float.ipynb <==
### Претворање на вредностите за SAPECoeficient, APECoeficient и  FPECoeficient од string во float

import numpy as np
import pandas as pd

df = pd.read_csv('dataset/Spenser_1_5_standardize_coefficients.tsv', delimiter='\t')
df

#### Претворање на вредностите за SAPECoeficient од string во float
Вредностите кои се обележани како ' ', ќе бидат заменети со <code>np.nan</code>.

sape = pd.Series([np.float(x) if x != ' ' else np.nan
                  for x in list(df['SAPECoeficient'])])

sape[:10]

df['SAPECoeficient'] = sape

#### Претворање на вредностите за APECoeficient од string во float
Вредностите кои се обележани како ' ', ќе бидат заменети со <code>np.nan</code>.

ape = pd.Series([np.float(x) if x != ' ' else np.nan
                 for x in list(df['APECoeficient'])])

ape[:10]

df['APECoeficient'] = ape

#### Претворање на вредностите за FPECoeficient од string во float
Вредностите кои се обележани како ' ', ќе бидат заменети со <code>np.nan</code>.

fpe = pd.Series([np.float(x) if x != ' ' else np.nan
                 for x in list(df['FPECoeficient'])])

fpe[:10]

df['FPECoeficient'] = fpe

#### Снимање на промените во нова датотека

df.to_csv('dataset/Spenser_1_6_str_to_float_coefficients.tsv',
          sep='\t', index=None)
