==> notebooks/01_clean_7_remove_V22_TRESHOLD_attributes.ipynb <==
### Отстранување на атрибутите V22 и TRESHOLD

import pandas as pd

df = pd.read_csv('dataset/Spenser_1_6_str_to_float_coefficients.tsv',
                 delimiter='\t')
df = df.drop(['V22', 'TRESHOLD'], axis=1)
df

df.to_csv('dataset/Spenser_1_7_remove_v22_treshold.tsv',
          sep='\t', index=None)
