==> notebooks/02_check_coefficient_values_consistency.ipynb <==
### Проверка дали вредностите за коефициентите $\in [0, 100]$

import pandas as pd

df = pd.read_csv('dataset/Spenser_1_7_remove_v22_treshold.tsv', sep='\t')
df.index = pd.RangeIndex(start=2, stop=len(df)+2)
df.index.name = 'XLS_RowNumber'
df.head()

cols = df.columns

#### Проверка за конзистентност на коефициентите

df_unc = pd.DataFrame(columns=cols)

for row in df.iterrows():
    if (row[1]['SAPECoeficient'] is not pd.np.nan and
        (row[1]['SAPECoeficient'] < 0 or row[1]['SAPECoeficient'] > 100)):
        df_unc = df_unc.append(pd.DataFrame(row[1]).transpose())
    elif (row[1]['APECoeficient'] is not pd.np.nan and
        (row[1]['APECoeficient'] < 0 or row[1]['APECoeficient'] > 100)):
        df_unc = df_unc.append(pd.DataFrame(row[1]).transpose())
    elif (row[1]['FPECoeficient'] is not pd.np.nan and
        (row[1]['FPECoeficient'] < 0 or row[1]['FPECoeficient'] > 100)):
        df_unc = df_unc.append(pd.DataFrame(row[1]).transpose())

df_unc.index.name = 'XLS_RowNumber'
df_unc

print(len(df_unc), 'редици со неконзистентни вредности за коефициентите')

#### Зачувување на редиците со неконзистентни коефициенти

df_unc.to_csv('subsets/unconsistent_coefficients.tsv', sep='\t')
