==> notebooks/02_check_consistency_in_coefficients_validation.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('dataset/Spenser_1_5_standardize_coefficients.tsv', sep='\t')

rows = list(df.iterrows())

def get_ape_coefficient(transaction, tid):
    for row in rows[tid:]:
        if row[1]['APECoeficient'] != ' ':
            return float(row[1]['APECoeficient'])


def check_consistency(transaction, tid):
    label = transaction['OK'] if transaction['OK'] != ' ' else transaction['NOTOK']
    ape = get_ape_coefficient(transaction, tid)
    sape = float(transaction['SAPECoeficient'])
    if sape == 0: # проверка за DivisionByZero exception
        return True
            
    diff = (1.0 - (ape/sape))
    
    if label == 'ok':
        if diff >= 0.15:
            return False
    else:
        if diff < 0.15:
            return False
    return True

inconsistencies = []
for row in rows:
    transaction = row[1]
    if transaction['OK'] == ' ' and transaction['NOTOK'] == ' ': # трансакцијата не ја евалуирал контролор, не е важна
        continue

    if not check_consistency(transaction, row[0]):
        inconsistencies.append(row[0]) 
    
print(inconsistencies)

Листата е празна, се' е во ред
