==> notebooks/02_clean_feature_vectors.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('dataset/Feature_vectors_by_customerID_and_month.tsv', sep='\t')
df.head()

no_records_wno_validation = len(df[df.No_validations == 0])
print(no_records_wno_validation)

За 30254 записи не постои валидација, односно атрибутот No_validations има вредност 0. Бидејќи не постои валидација, ова повлекува дека и атрибутите Mean_APE, Squared_error_SAPE_APE и Day_validation ќе имаат np.nan вредности.
Nan вредностите на овие атрибути ќе бидат заменети со -1. 

mean_ape_no_nan = np.zeros(len(df))
sqrd_err_no_nan = np.zeros(len(df))
day_of_validation_no_nan = np.zeros(len(df))
    
for i in range(len(df)):
    mean_ape = df['Mean_APE'][i]
    sqrd_err = df['Squared_error_SAPE_APE'][i]
    day_of_validation = df['Day_validation'][i]
    
    mean_ape_no_nan[i] = -1 if np.isnan(mean_ape) else mean_ape
    sqrd_err_no_nan[i] = -1 if np.isnan(sqrd_err) else sqrd_err
    day_of_validation_no_nan[i] = -1 if np.isnan(day_of_validation) else day_of_validation

df.Mean_APE = pd.Series(mean_ape_no_nan, dtype=float)
df.Squared_error_SAPE_APE = pd.Series(sqrd_err_no_nan, dtype=float)
df.Day_validation = pd.Series(day_of_validation_no_nan, dtype=int)

df[(df.No_executed_visits == 0) & (df.No_adhoc_visits == 0) & (~np.isnan(df.Mean_SAPE))]

df[(df.No_executed_visits == 0) & (df.No_adhoc_visits == 0) & (~np.isnan(df.Mean_FPE))]

Отстранување на редиците за кои што атрибутите No_executed_visits и No_adhoc_visits имаат вредност 0. Ова значи дека во даден објект (CustomerID), за даден месец (Month) не е извршена ниту планирана, ниту адхок посета. 

Постојат само 2 записи за кои што овие два атрибути имаат вредност 0, а вредноста на Mean_SAPE или Mean_FPE е различна од np.nan.
Овие записи се најверојатно грешка, бидејќи во тој месец во објектот не е направена никаква посета од страна на комерцијалист, а постојат коефициенти за евалуација на анкета.

По ова чистење, може да се отстрани и атрибутот No_validations (вредности 0 и 1), бидејќи секогаш кога овој атрибут има вредност 0, Day_validation има вредност -1, а секогаш кога овој атрибут има вредност 1, Day_validation е позитивен цел број. 

df = df.drop(df.index[(df.No_executed_visits == 0) & (df.No_adhoc_visits == 0)], axis=0)
df = df.drop(['No_validations'], axis=1)

strange_customers = df[(df.No_executed_visits > 0) & (np.isnan(df.Mean_SAPE))]
strange_customers.to_csv('subsets/Customers_w_executed_visits_and_no_sape.tsv', sep='\t', index=None)
len(strange_customers)

Претходните записи се неконзистентни. За даден објект и месец, забележано е дека посета (или повеќе посети) од страна на комерцијалист е извршена, но не е забележан никаков SAPE коефициент.

Претходните 65 записи ќе бидат отстранети бидејќи немаат никакво значење.

df = df.drop(df.index[(df.No_executed_visits > 0) & (np.isnan(df.Mean_SAPE))], axis=0)
len(df)

not all(map(np.isnan, df.Mean_FPE))  # нема ниту еден запис со MEAN_FPE = np.nan

not all(map(np.isnan, df.Mean_SAPE)) # нема ниту еден запис со MEAN_SAPE = np.nan

df.to_csv('dataset/Feature_vectors_customer_month_cleaned.tsv', sep='\t', index=None)
