==> notebooks/02_extract_performance_evaluation_subsets.ipynb <==
### Извлекување на подмножества од сите можни комбинации за PerformanceEvaluationPlanned, PerformanceEvaluationExecuted и PerformanceEvaluationAdHoc

import pandas as pd

df = pd.read_csv('dataset/Spenser_1_7_remove_v22_treshold.tsv', sep='\t')
df.index = pd.RangeIndex(start=2, stop=len(df)+2)
df.index.name = 'XLS_RowNumber'
df.head()

#### Кои се можни вредности за PerformanceEvaluationPlanned, PerformanceEvaluationExecuted и PerformanceEvaluationAdHoc поединечно

set(df['PerformanceEvaluationPlanned'])

set(df['PerformanceEvaluationExecuted'])

set(df['PerformanceEvaluationAdHoc'])

#### Сите можни комбинации за (Planned, Executed, AdHoc)

s = set(tuple([i[3], i[4], i[5]])
        for i in list(df.values))

s = sorted(list(s))

for i in s:
    print(i)

#### Извлекување и снимање на подмножества за сите можни комбинации за (Planned, Executed, AdHoc)

for i in s:
    p, e, a = i
    t = df[(df.PerformanceEvaluationPlanned == p) &
           (df.PerformanceEvaluationExecuted == e) &
           (df.PerformanceEvaluationAdHoc == a)]
    p = 'n' if p == ' ' else p
    e = 'n' if e == ' ' else e
    a = 'n' if a == ' ' else a
    t.to_csv('subsets/PerformanceEvaluation/{}_{}_{}.tsv'.format(p, e, a), sep='\t')
