==> notebooks/02_extract_possibly_redundant_data.ipynb <==
### Извлекување на податоци кои што е можно да бидат редундантни. Извлечени се следниве подмножества на податоци:
    - Објекти за кои што не е планирана ниту една посета
    - Објекти за кои што не е извршена ниту една планирана посета
    - Објекти за кои што не е извршена ниту една валидација
    - Објекти за кои што не е извршена ниту една планирана посета и ниту една валидација
    - Објекти за кои што не е извршена ниту една посета (планирана или адхок) и ниту една валидација

import pandas as pd
import numpy as np

df_features = pd.read_csv('dataset/Feature_vectors_by_customerID.tsv', sep='\t')
df = pd.read_csv('dataset/Spenser_1_7_remove_v22_treshold.tsv', sep='\t')

no_planned_visits = df_features[df_features['No_planned_visits'] == 0]
objects_w_no_planned_visits = no_planned_visits.CustomerID  #објекти за кои не е извршена ниту една планирана посета

print('Постојат {} објекти за кои што не е планирана ниту една посета од страна на комерцијалист'.format(len(objects_w_no_planned_visits)))

objects_w_no_planned_visits.to_csv('subsets/PossiblyRedundantData/objects_w_no_planned_visits.tsv', sep='\t', index=None)

no_executed_visits = df_features[df_features['No_executed_visits'] == 0]
objects_w_no_executed_visits = no_executed_visits.CustomerID  #објекти за кои не е извршена ниту една планирана посета

print('Постојат {} објекти за кои што не е извршена ниту една планирана посета од страна на комерцијалист'.format(len(objects_with_no_executed_visits)))

objects_w_no_executed_visits.to_csv('subsets/PossiblyRedundantData/objects_w_no_executed_visits.tsv', sep='\t', index=None)

no_validations = df_features[df_features['No_validations'] == 0]
objects_w_no_validations = no_validations.CustomerID  #објекти за кои што не е извршена ниту една валидација од страна на контролор

print('Постојат {} објекти за кои што не е извршена ниту една валидација од страна на контролор'.format(len(objects_w_no_validations)))

objects_w_no_validations.to_csv('subsets/PossiblyRedundantData/objects_w_no_validations.tsv', sep='\t', index=None)

no_validations_nor_executed = df_features[(df_features['No_validations'] == 0) &
                                          (df_features['No_executed_visits'] == 0)]

objects_w_no_val_nor_executed = no_validations_nor_executed.CustomerID  #објекти за кои што не е извршена ниту една планирана посета од страна на комерцијалист и не е извршена ниту една валидација од страна на контролор

print('Постојат {} објекти за кои што не е извршена ниту една планирана посета од страна на комерцијалист и ' \
      'не е извршена ниту една валидација од страна на контролор'.format(len(objects_w_no_val_nor_executed)))

objects_w_no_val_nor_executed.to_csv('subsets/PossiblyRedundantData/objects_w_no_val_nor_executed.tsv', sep='\t', index=None)

no_valid_nor_executed_nor_adhoc = df_features[(df_features['No_validations'] == 0) 
                                              & (df_features['No_executed_visits'] == 0)
                                            & (df_features['No_adhoc_visits'] == 0)]

objects_w_no_val_nor_executed_nor_adhoc = no_valid_nor_executed_nor_adhoc.CustomerID   #објекти за кои што не е извршена ниту една планирана ниту адхок посета од страна на комерцијалист и не е извршена ниту една валидација од страна на контролор

print('Постојат {} oбјекти за кои што не е извршена ниту една планирана, ниту адхок посета од страна на комерцијалист и не е извршена ниту една валидација од страна на контролор'.format(len(objects_w_no_val_nor_executed_nor_adhoc)))

objects_w_no_val_nor_executed_nor_adhoc.to_csv('subsets/PossiblyRedundantData/objects_w_no_val_nor_executed_nor_adhoc.tsv', sep='\t', index=None)
