==> notebooks/02_visits_per_445_month.ipynb <==
### Проверка колку пати е посетена секоја продавница во 4-4-5 месец

import math
import numpy as np
import pandas as pd

df = pd.read_csv('dataset/Spenser_1_5_standardize_coefficients.tsv',
                 delimiter='\t')
df

#### ...

cols = ['SAPECCSalesGroup', 'TERRID', 'CustomerID', 'month', 'Visits']

df_new = pd.DataFrame(columns=cols)

for team in set(df['SAPECCSalesGroup']):
    t1 = df[df['SAPECCSalesGroup'] == team]
    
    for employee in set(t1['TERRID']):
        t2 = t1[t1['TERRID'] == employee]
    
        for customer in set(t2['CustomerID']):
            t3 = t2[t2['CustomerID'] == customer]

            for month in set(t3['month']):
                s = set(t3[t3['month'] == month]['APECoeficient'])
                visits = len(s) - (' ' in s)

                df_new = df_new.append(pd.DataFrame([[team, employee, customer,
                                                      month, visits]], columns=cols), )

df_new

#### Подредување

df_new = df_new.sort_values(cols[:-1])

df_new

#### Снимање на обработените податоци

df_new.to_csv('dataset/Spenser_2_visits_per_445_month.tsv',
              sep='\t', index=None)
