==> notebooks/03_m3_beta_distr_01a_data_extraction.ipynb <==
### Модел 2: Моделирање со помош на Beta распределба
#### Припрема на податоците за обработка

import numpy as np
import os
import pandas as pd

#### Читање на влезните податоци

df = pd.read_csv('../data/dataset/Spenser_1_8_new_format_on_PE_markers.tsv', sep='\t')

# Отстранување на редиците каде нема ниту една вредност за 'SAPECoeficient', 'APECoeficient', 'FPECoeficient'
valid_indices = df.index[(df.notna()['SAPECoeficient'] | \
                          df.notna()['APECoeficient'] | \
                          df.notna()['FPECoeficient'])]
df = df.iloc[valid_indices]

# Приказ
df[::len(df)//10]

print(f'Остануваат {len(df)} валидни ставки за работа.')

#### Проверка дали има подмножество ('CustomerID', 'month') за кое е извршена повеќе од 1 контролна посета

#### <font style="color: #FF0000">Спора операција!!!</font>

for cid in set(df['CustomerID']):
    for month in set(df[df['CustomerID'] == cid]['month']):
        sub_df = df[(df['CustomerID'] == cid) & (df['month'] == month)]
        
        # Земи ги сите ad-hoc посети за ова подмножество ('CustomerID', 'month')
        ad_hoc_df = sub_df[(sub_df['PerformanceEvaluationPlanned'] == 0) & \
                           (sub_df['PerformanceEvaluationExecuted'] == 1)]

        # Земи ги сите планирани посети за ова подмножество ('CustomerID', 'month')
        planned_df = sub_df[(sub_df['PerformanceEvaluationPlanned'] == 1) & \
                            (sub_df['PerformanceEvaluationExecuted'] == 1)]

        # Земи ги сите контролни посети за ова подмножество ('CustomerID', 'month')
        external_df = sub_df[sub_df['ExternalEvaluation'] == 1]
        
        if len(external_df) > 1:
            display(sub_df)

Не е извршена повеќе од 1 контролна посета во подмножествата ('CustomerID', 'month').

#### Собирање на примероци

ad_hoc_pre   = list()  # пред да се случи ExternalEvaluation
ad_hoc_pst   = list()  # после ExternalEvaluation 
planned_pre  = list()  # пред да се случи ExternalEvaluation
planned_pst  = list()  # после ExternalEvaluation 
external_val = list()  # вледностите на ExternalEvaluation

for cid in set(df['CustomerID']):
    for month in set(df[df['CustomerID'] == cid]['month']):
        sub_df = df[(df['CustomerID'] == cid) & (df['month'] == month)]
        
        # Земи ги сите ad-hoc посети за ова подмножество ('CustomerID', 'month')
        ad_hoc_df = sub_df[(sub_df['PerformanceEvaluationPlanned'] == 0) & \
                           (sub_df['PerformanceEvaluationExecuted'] == 1)]

        # Земи ги сите планирани посети за ова подмножество ('CustomerID', 'month')
        planned_df = sub_df[(sub_df['PerformanceEvaluationPlanned'] == 1) & \
                            (sub_df['PerformanceEvaluationExecuted'] == 1)]

        # Земи ги сите контролни посети за ова подмножество ('CustomerID', 'month')
        external_df = sub_df[sub_df['ExternalEvaluation'] == 1]
        
        if len(external_df) == 0:  # не постои контролна посета
            ad_hoc_pre  += list(ad_hoc_df['SAPECoeficient'].values)  # .dropna().values
            planned_pre += list(planned_df['SAPECoeficient'].values)  # .dropna().values
        else:
            external_val += list(external_df['APECoeficient'].values)

            for item in planned_df.iterrows():
                if item[1]['Calendarday'] < external_df.iloc[0]['Calendarday']:
                    planned_pre.append(item[1]['SAPECoeficient'])
                elif item[1]['Calendarday'] > external_df.iloc[0]['Calendarday']:
                    planned_pst.append(item[1]['SAPECoeficient'])

            for item in ad_hoc_df.iterrows():
                if item[1]['Calendarday'] < external_df.iloc[0]['Calendarday']:
                    ad_hoc_pre.append(item[1]['SAPECoeficient'])
                elif item[1]['Calendarday'] > external_df.iloc[0]['Calendarday']:
                    ad_hoc_pst.append(item[1]['SAPECoeficient'])

#### Запишување на излезните податоци во npy формат

data_out = '../data/dataset/03_m3_beta_distr_01a_data_extraction'

if not os.path.exists(data_out):
    os.makedirs(data_out)
    
ad_hoc_pre_arr = np.array(ad_hoc_pre)
ad_hoc_pre_arr = ad_hoc_pre_arr[~np.isnan(ad_hoc_pre_arr)]
if not os.path.exists(f'{data_out}/ad_hoc_pre'):
    np.save(f'{data_out}/ad_hoc_pre', ad_hoc_pre_arr)

ad_hoc_pst_arr = np.array(ad_hoc_pst)
ad_hoc_pst_arr = ad_hoc_pst_arr[~np.isnan(ad_hoc_pst_arr)]
if not os.path.exists(f'{data_out}/ad_hoc_pst'):
    np.save(f'{data_out}/ad_hoc_pst', ad_hoc_pst_arr)

ad_hoc_all_arr = np.concatenate((ad_hoc_pre_arr, ad_hoc_pst_arr))
if not os.path.exists(f'{data_out}/ad_hoc_all'):
    np.save(f'{data_out}/ad_hoc_all', ad_hoc_all_arr)

planned_pre_arr = np.array(planned_pre)
planned_pre_arr = planned_pre_arr[~np.isnan(planned_pre_arr)]
if not os.path.exists(f'{data_out}/planned_pre'):
    np.save(f'{data_out}/planned_pre', planned_pre_arr)

planned_pst_arr = np.array(planned_pst)
planned_pst_arr = planned_pst_arr[~np.isnan(planned_pst_arr)]
if not os.path.exists(f'{data_out}/planned_pst'):
    np.save(f'{data_out}/planned_pst', planned_pst_arr)

planned_all_arr = np.concatenate((planned_pre_arr, planned_pst_arr))
if not os.path.exists(f'{data_out}/planned_all'):
    np.save(f'{data_out}/planned_all', planned_all_arr)

external_arr = np.array(external_val)
external_arr = external_arr[~np.isnan(external_arr)]
if not os.path.exists(f'{data_out}/external'):
    np.save(f'{data_out}/external', external_arr)
